- warning fixes; - more cleaning
git-svn-id: http://svn.miranda-ng.org/main/trunk@17168 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ plugins/TabSRMM/src/msgdialog.cpp
@@ -1334,7 +1334,7 @@ INT_PTR CALLBACK DlgProcMessage(HWND hwndDlg, UINT uMsg, WPARAM wParam, LPARAM l
 		dat->cache->updateStats(TSessionStats::INIT_TIMER);
 		if (dat->hContact) {
 			FindFirstEvent(dat);
-			dat->nMax = dat->cache->getMaxMessageLength();
+			dat->nMax = (int)dat->cache->getMaxMessageLength();
 		}
 		LoadContactAvatar(dat);
 		SendMessage(hwndDlg, DM_OPTIONSAPPLIED, 0, 0);
